requirements.txt
alembic==1.0.7
autopep8==1.4.3
blinker==1.4
Click==7.0
Flask==1.0.2
boto3==1.9.171
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pep8==1.7.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
WTForms==2.2.1
# requirements for Heroku
psycopg2==2.8.2
gunicorn==19.9.0
# for elastic search
urllib3==1.25.2
elasticsearch==7.0.1
